Transceiver52M: Add RACH flood testing option
Allow flag passing to the transceiver constructor with a single RACH
flood flag. This test mode sets the correlation type to RACH on all
slots and disables selective burst detection - detection is always
true. Demod uses a random symbol timing offset because the return TOA
value would typically be zero in this case.

Note that when 4 samples-per-symbol are used on the downlink, this
option has only a minor role in load testing because the receive path
is not oversampled and processing is dominated higher rate on the
transmit side.

Signed-off-by: Thomas Tsou <tom@tsou.cc>