Transceiver52M: Decode SCH and adjust GSM clock
When in MS acquisition mode, attempt to decode SCH and establish the BTS
frame timing. Lock the transceiver GSM clock to the BTS by adjusting the
clock value by the measured burst-SCH offset.

Add tracking state, TRX_MODE_MS_TRACKING, which continues to detect and
decode the SCH with timing tracking, but only on SCH poitions within the
51 multiframe.

Signed-off-by: Thomas Tsou <tom@tsou.cc>