Transceiver52M: Use multiple SCH correlation ranges
Setup two SCH detection ranges - full search and narrow search. Full
search correlates the full burst length to find for SCH detection.
Narrow uses a reduced search range, which is the same as that used for
RACH detection.

Narrow searching is enabled after the SCH is successfully decoded and
the MS is locked to the frame timing.

Signed-off-by: Thomas Tsou <tom@tsou.cc>